fix: user server

+++ prisma/migrations/20241111174425_init/migration.sql
@@ -3,9 +3,9 @@ CREATE TABLE "User" (
     "id" TEXT NOT NULL,
     "login" TEXT NOT NULL,
     "password" TEXT NOT NULL,
-    "version" INTEGER NOT NULL DEFAULT 1,
-    "createdAt" INTEGER NOT NULL,
-    "updatedAt" INTEGER NOT NULL,
+    "version" BIGINT NOT NULL DEFAULT 1,
+    "createdAt" BIGINT NOT NULL,
+    "updatedAt" BIGINT NOT NULL,
 
     CONSTRAINT "User_pkey" PRIMARY KEY ("id")
 );
